requirements.txt
numpy
tensorflow
opencv-python
requests
paho-mqtt